Refactoring following plan in refactor_plan.md. Phase 1 done
