+++ src/aind_session_utils/sources/__init__.py
@@ -1,0 +1,10 @@
+"""Data source modules for aind-session-utils."""
+
+SOURCE_MODULES = {
+    "docdb": "aind_session_utils.sources.docdb",
+    "dts": "aind_session_utils.sources.dts",
+    "codeocean": "aind_session_utils.sources.codeocean",
+    "watchdog": "aind_session_utils.sources.watchdog",
+    "manifests": "aind_session_utils.sources.manifests",
+    "rig_logs": "aind_session_utils.sources.rig_logs",
+}
